Update script.js to work with APi
Had to get rid of Google sheets and the API Key for that and I had to sign up and create an account to get a US Department of Education API key. This uses US News numbers and automatically updates whenever the numbers change. The issue right now is that it only has certain colleges and it is laggy because of the number of schools.