todays description
Going to try to solve the animations and the css tommorow. Right now all the school names are in white so you can't see it but that is an easy fix. Today I did research for the google maps and I added the mascots to come up when the user searches. Some problems im having is how to center the search boxes but Ill work on that next week, this week im going to get the functionality down